types: move TypeInner and UIntType to inner.rs module
code move only
